refactor: representar dinero como enteros

+++ app/src/main/java/com/ivanmadrid/vehiclecontrolapp/domain/model/Novelty.kt
@@ -9,5 +9,5 @@ data class Novelty(
     val priority: NoveltyPriority,
     val affectsIncome: Boolean,
     val incomeAdjustmentType: IncomeAdjustmentType?,
-    val adjustedIncomeAmount: Double?,
+    val adjustedIncomeAmount: Long?,
 )
