+++ app/src/main/java/com/ivanmadrid/vehiclecontrolapp/domain/model/Vehicle.kt
@@ -8,5 +8,5 @@ data class Vehicle(
     val type: VehicleType,
     val status: String,
     val currentDriver: String?,
-    val dailyFixedIncome: Double?,
+    val dailyFixedIncome: Long?,
 )
